app/requirements.txt
flask==2.0.1
numpy
requests
scikit-learn
uWSGI
mecab-python3
unidic-lite
wordcloud
firebase-admin